# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A high-performance Go API gateway for interacting with multiple blockchain networks through a unified interface.\n\nThis API provides simplified access to various blockchain RPC endpoints through a single, consistent interface.\nIt abstracts away the differences between blockchain implementations, allowing developers to focus on building their applications.",
        "title": "Blockchain Gateway API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/dvictor357/blockchain-gateway",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/batch": {
            "post": {
                "description": "Execute multiple RPC queries across different blockchain networks in a single request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chains"
                ],
                "summary": "Execute Batch RPC Queries",
                "parameters": [
                    {
                        "description": "Batch RPC Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/api.SwaggerRPCRequest"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/api.SwaggerRPCResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/chains": {
            "get": {
                "description": "Get a list of all blockchain networks supported by this gateway",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chains"
                ],
                "summary": "List Supported Blockchains",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ChainsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/chains/{chain}/address/{address}/balance": {
            "get": {
                "description": "Get the native token balance for a specific address on a blockchain",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chains"
                ],
                "summary": "Get Account Balance",
                "parameters": [
                    {
                        "enum": [
                            "ethereum",
                            "bitcoin",
                            "polygon"
                        ],
                        "type": "string",
                        "description": "Blockchain name",
                        "name": "chain",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Wallet address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerBalance"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/chains/{chain}/address/{address}/nonce": {
            "get": {
                "description": "Get the number of transactions for a specific address on a blockchain",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chains"
                ],
                "summary": "Get Transaction Count",
                "parameters": [
                    {
                        "enum": [
                            "ethereum",
                            "bitcoin",
                            "polygon"
                        ],
                        "type": "string",
                        "description": "Blockchain name",
                        "name": "chain",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Wallet address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.TransactionCountResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/chains/{chain}/block/latest": {
            "get": {
                "description": "Get information about the latest block on a blockchain",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chains"
                ],
                "summary": "Get Latest Block",
                "parameters": [
                    {
                        "enum": [
                            "ethereum",
                            "bitcoin",
                            "polygon"
                        ],
                        "type": "string",
                        "description": "Blockchain name",
                        "name": "chain",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerBlockInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/chains/{chain}/gas-price": {
            "get": {
                "description": "Get the current gas price for EVM-compatible blockchains",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chains"
                ],
                "summary": "Get Gas Price",
                "parameters": [
                    {
                        "enum": [
                            "ethereum",
                            "polygon"
                        ],
                        "type": "string",
                        "description": "Blockchain name",
                        "name": "chain",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GasPriceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/chains/{chain}/query": {
            "post": {
                "description": "Execute a JSON-RPC query on a specific blockchain network",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chains"
                ],
                "summary": "Execute RPC Query",
                "parameters": [
                    {
                        "enum": [
                            "ethereum",
                            "bitcoin",
                            "polygon"
                        ],
                        "type": "string",
                        "description": "Blockchain name",
                        "name": "chain",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "RPC Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerRPCRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerRPCResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "504": {
                        "description": "Gateway Timeout",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/chains/{chain}/tx/{hash}": {
            "get": {
                "description": "Get detailed information about a specific transaction by its hash",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chains"
                ],
                "summary": "Get Transaction Details",
                "parameters": [
                    {
                        "enum": [
                            "ethereum",
                            "bitcoin",
                            "polygon"
                        ],
                        "type": "string",
                        "description": "Blockchain name",
                        "name": "chain",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Transaction hash",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerTransactionInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.GasPriceResponse": {
            "type": "object",
            "properties": {
                "chain": {
                    "type": "string",
                    "example": "ethereum"
                },
                "gas_price": {
                    "type": "string",
                    "example": "20000000000"
                },
                "gas_price_hex": {
                    "type": "string",
                    "example": "0x4a817c800"
                }
            }
        },
...
        "api.SwaggerErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/validation.ValidationError"
                    }
                },
                "error": {
                    "type": "string",
                    "example": "Validation failed"
                }
            }
        },
        "api.SwaggerRPCRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "jsonrpc": {
                    "type": "string",
                    "example": "2.0"
                },
                "method": {
                    "type": "string",
                    "example": "eth_blockNumber"
                },
                "params": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[]"
                    ]
                }
            }
        },
        "api.SwaggerRPCResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/api.RPCError"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "jsonrpc": {
                    "type": "string",
                    "example": "2.0"
                },
                "result": {
                    "type": "string",
                    "example": "0x1234567"
                }
            }
        },
...
    },
...
}